add update time

+++ routers/upload.py
@@ -1,4 +1,5 @@
-import json
+from os.path import exists
+from json import dump, load
 from datetime import datetime
 
 from fastapi import APIRouter, Cookie, Query, UploadFile, File
@@ -23,8 +24,14 @@ async def upload_file(path: Optional[str] = Query(None), data: UploadFile = File
         with open(f"temp/{path}/{data.filename}", "wb") as uploaded_file:
             uploaded_file.write(await data.read())
         date = datetime.now().strftime("%d-%m-%y %H:%M")
-        with open(f"temp/{path}/{data.filename}.meta", "w") as meta_file:
-            json.dump({"create": str(date), "modif": str(date)}, meta_file)
+        if exists(f"temp/{path}/{data.filename}.meta"):
+            with open(f"temp/{path}/{data.filename}.meta", "r") as meta_file:
+                meta_data = load(meta_file)
+            with open(f"temp/{path}/{data.filename}.meta", "w") as meta_file:
+                dump({"create": str(meta_data["create"]), "modif": str(date)}, meta_file)
+        else:
+            with open(f"temp/{path}/{data.filename}.meta", "w") as meta_file:
+                dump({"create": str(date), "modif": str(date)}, meta_file)
         return RedirectResponse(f"/{path}", status_code=302)
     except Exception as er:
         print(er)
